Add initial test code for new traces.

+++ Src/Sarif.PatternMatcher/SearchSkimmer.cs
@@ -1437,7 +1437,7 @@ namespace Microsoft.CodeAnalysis.Sarif.PatternMatcher
             }
         }
 
-        private bool DoesTargetFileExceedSizeLimits(long fileLength, int maxFileSize)
+        private bool DoesTargetFileExceedSizeLimits(long fileLength, long maxFileSize)
         {
             // Ensure that the byte of the file does not exceed the limit set by the
             // file-size-in-kilobytes command line argument, which defaults to ~10MB.
